Finished user managemet

+++ apps/gateway/src/app/users/users.controller.ts
@@ -1,31 +1,49 @@
-import { Body, Controller, Delete, Get, Param, Post, UseGuards } from '@nestjs/common'
-import { eUserRole, GATEWAY_ROUTES } from '@sentinel-supreme/shared'
-import { RegisterDto } from '@sentinel-supreme/shared/server'
+import {
+	BadRequestException,
+	Body,
+	Controller,
+	Delete,
+	Get,
+	Param,
+	Post,
+	Put,
+	UseGuards
+} from '@nestjs/common'
+import { eUserRole, GATEWAY_ROUTES, iUser } from '@sentinel-supreme/shared'
+import { RegisterDto, UpdateDto } from '@sentinel-supreme/shared/server'
+import { GetUser } from '../auth/decorators/get-user.decorator'
 import { Roles } from '../auth/decorators/roles.decorator'
 import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard'
 import { RolesGuard } from '../auth/guards/roles.guard'
 import { UsersService } from './users.service'
 
 @Controller(GATEWAY_ROUTES.USERS)
+@UseGuards(JwtAuthGuard, RolesGuard)
+@Roles(eUserRole.ADMIN)
 export class UsersController {
 	constructor(private readonly usersService: UsersService) {}
 
-	@Post(GATEWAY_ROUTES.REGISTER)
-	async register(@Body() body: RegisterDto) {
-		return this.usersService.create(body.email, body.password)
+	@Post()
+	async createUser(@Body() body: RegisterDto) {
+		return this.usersService.create(body.email, body.password, body.role)
 	}
 
 	@Get()
-	@UseGuards(JwtAuthGuard, RolesGuard)
-	@Roles(eUserRole.ADMIN)
-	async getAllUsers() {
+	async getAllUsers(): Promise<iUser[]> {
 		return this.usersService.getAll()
 	}
 
 	@Delete(':id')
-	@UseGuards(JwtAuthGuard, RolesGuard)
-	@Roles(eUserRole.ADMIN)
-	async deleteUser(@Param('id') id: string) {
+	async deleteUser(@Param('id') id: string, @GetUser('userId') requesterId: string) {
+		if (id === requesterId) {
+			throw new BadRequestException('Self-deletion is not allowed')
+		}
+
 		return this.usersService.deleteById(id)
+	}
+
+	@Put(':id')
+	async updateUser(@Param('id') id: string, @Body() body: UpdateDto) {
+		return this.usersService.update(id, body)
 	}
 }
